### 2023-01-25 #74 @skrawcz
Hamilton
Submitting Author: Stefan Krawczyk (@skrawcz)
Package Name: Hamilton (sf-hamilton on pypi)
One-Line Description of Package: A general purpose micro-framework for defining dataflows.
Repository Link (if existing):   https://github.com/stitchfix/hamilton

---

Hamilton is a general purpose micro-framework for creating dataflows from python functions! Specifically, Hamilton defines a novel paradigm, that allows you to specify a flow of (delayed) execution, that forms a Directed Acyclic Graph (DAG). It was originally built to solve the challenges in wrangling and maintaining production code to create wide (1000+) column dataframes, but has been extended to enable modeling any python object generation. Core to the design of Hamilton is a clear mapping of function name to dataflow output. That is, Hamilton forces a declarative paradigm expressed through writing python functions, and aims for DAG clarity, low code upkeep costs, ease of modification, with always unit testable and naturally documentable code.

*data munging*
Hamilton was built for a team to manage their time-series forecasting feature engineering. So it's design goal was to help data science teams maintain data munging code well.

*reproducibility*
Core to reproducibility is sharing code. Most researchers only share data, not their code. We believe that with Hamilton, one could more easily share their implementation and in a standardized way that is approachable to a broad audience.

*data extraction*
Kind of unsure here. But Hamilton helps you structure and "orchestrate" the code that does extraction.

*data retrieval*
Kind of unsure here. But Hamilton helps you structure and "orchestrate" the code that does retrieval.

Anyone doing any data transformations in python.

Scientific applications: time-series forecasting, any machine learning, any work that involves executing a dataflow.

None that the author is aware of.

N/A
