### 2023-01-25 90f69a9
Adds one more assert to test
