### 2023-01-25 #74 @NickleDave
Hi @skrawcz welcome to pyOpenSci! Thank you for your presubmission inquiry.

I'll help you figure out if the package is in scope, and if we should move to a full submission, at which time I would find an editor.

The categories you have indicated look right to me.

A couple of questions:
* I have not used data orchestration tools extensively but my understanding is that Hamilton is similar to tools like airflow, metaflow, dagster, luigi, etc., as one of your core contributors comments [here](https://atlan.com/open-source-data-orchestration-tools/). I'm a bit surprised at the statement that there's no other Python packages that accomplish similar things.

Could you please clarify how Hamilton relates to those other tools?

* The most similar package I'm aware of in the open science space is [snakemake](https://snakemake.readthedocs.io/en/stable/index.html). If it helps, here's [results from clicking on the snakemake "topic" on GitHub](https://github.com/topics/snakemake) many examples of where researchers have shared their code that relies on the package. Again, I don't know data orchestration tools super well, but if there's signficant overlap then I'm not sure this would be quite in scope for us.

Can you talk about overlap with snakemake? If you have examples where researchers are already using Hamilton, that would help.

* And please just say a little generally about what the Hamilton authors are hoping to achieve by a pyOpenSci review.

To be clear, I fully agree with you that Hamilton is a tool that scientists could potentially use for munging and reproducibility. But our focus is on helping authors make sure their packages are more usable by scientists. I get the impression that development of Hamilton is already quite mature (and already supported by a tech company 🙂 ) so I'm not sure how beneficial it would be to you all to go through review.

Thank you!
