### 2023-01-26 #74 @skrawcz
> I have not used data orchestration tools extensively but my understanding is that Hamilton is similar to tools like airflow, metaflow, dagster, luigi, etc., as one of your core contributors comments [here](https://atlan.com/open-source-data-orchestration-tools/). I'm a bit surprised at the statement that there's no other Python packages that accomplish similar things.
> Could you please clarify how Hamilton relates to those other tools?

Sure.

Hamilton does not replace airflow, metaflow, dagster, luigi, etc. They focus on the "macro" scheduling problem, and are systems that require state to be managed. Hamilton focuses on the "micro", i.e. what people do within the step of an airflow, metaflow, dagster, luigi, etc. task. Hamilton replaces lines of logic with functions and tries to make that part of a code base testable, documentation friendly, and maintainable, which is not the goal of those other systems.

For example, here's a blog post showing [Hamilton + Metaflow](https://outerbounds.com/blog/developing-scalable-feature-engineering-dags/) - Hamilton helps with the feature engineering task, and metaflow does the macro orchestration.

Other differences:
-  the orchestration systems mentioned are generally only useful when data is too big to fit in memory. Hamilton is useful from day 1 of a project -- it helps organize the code to be run.
- Hamilton is not opinionated with respect to execution. The user needs to manage the python environment and place of execution for Hamilton code to run. That's why you can run Hamilton in a webservice, jupyter notebook, airflow, etc., but can't with those other systems.

> The most similar package I'm aware of in the open science space is [snakemake](https://snakemake.readthedocs.io/en/stable/index.html). If it helps, here's [results from clicking on the snakemake "topic" on GitHub](https://github.com/topics/snakemake) many examples of where researchers have shared their code that relies on the package. Again, I don't know data orchestration tools super well, but if there's signficant overlap then I'm not sure this would be quite in scope for us.
> Can you talk about overlap with snakemake? If you have examples where researchers are already using Hamilton, that would help.

From snakemake:
> The Snakemake workflow management system is a tool to create reproducible and scalable data analyses. Workflows are described via a human readable, Python based language. They can be seamlessly scaled to server, cluster, grid and cloud environments, without the need to modify the workflow definition. Finally, Snakemake workflows can entail a description of required software, which will be automatically deployed to any execution environment.

Snakemake is basically an orchestration system it sounds like. Hamilton is much much lighter-weight and only focused on pure python. Given the feature set snakemake has, and if that's the bar for reproducibility, then I don't think Hamilton meets it. I'll retract that category.

> And please just say a little generally about what the Hamilton authors are hoping to achieve by a pyOpenSci review.
> To be clear, I fully agree with you that Hamilton is a tool that scientists could potentially use for munging and reproducibility. But our focus is on helping authors make sure their packages are more usable by scientists. I get the impression that development of Hamilton is already quite mature (and already supported by a tech company 🙂 ) so I'm not sure how beneficial it would be to you all to go through review.

(1) I saw pandera on the list, and I think it's a good tool for a scientist to know about (Hamilton supports integration with it), and thus thought to myself that Hamilton would be a fit here.
(2) I genuinely think it's a good tool for anyone doing python data transform work (e.g. pandas), back when I was studying research code wasn't the best quality. I think tools like Hamilton & Pandera can help teams ensure their software projects live for much longer than their PhD.
(3) we have people at labs picking Hamilton up, e.g. at https://www.pnnl.gov/, and so I see a fit with scientists, and thus part of applying is to learn/understand how can we make it better and easier for scientists to pick up and use Hamilton?

Thanks for the questions -- let me know what I can clarify/you want to diver deeper on.
