### 2023-01-26 #74 @NickleDave
Thank you @skrawcz, that's very helpful.

I think I understand that Hamilton is a pure Python way to do feature engineering (in brief).

I'm glad you mentioned pandera, it did also occur to me they might work well together.

If there are any public repositories from nat'l labs like PNNL that would provide examples of using Hamilton in the wild, that could also help us see its application to open science.

I hear you that one of your goals is to better understand how to reach this community of users. So your reasons for wanting review make sense to me and make me feel like this could be in scope.

I am discussing with the executive director and other editors. Please let me get back to you with any further questions or a decision by Monday at the latest.
