### 2023-01-26 #74 @skrawcz
> If there are any public repositories from nat'l labs like PNNL that would provide examples of using Hamilton in the wild, that could also help us see its application to open science.

Sure -- here's what I've found in my notes:
example transforms:
https://github.com/IMMM-SFA/naturf/blob/feature/nodes/naturf/nodes.py
how it's run:
https://github.com/IMMM-SFA/naturf/blob/feature/nodes/naturf/driver.py
