### 2023-01-30 #74 @NickleDave
Hi again @skrawcz, thank you for providing that example. It's very helpful to see.

We have decided that yes we will review the package as it is in scope.

Some context on the decision for you, and us for future context, and transparency: as I noted above, we see that Hamilton has already had support for its development, and there is a proceedings paper, although a publication review is not the same as our software review. One of our goals is to provide resources to packages that have not yet enjoyed this kind of support. But it is also within our scope to help build consistency across the whole scientific Python ecosystem. You have clearly shown that (1) you are interested in participating in this process as an author and (2) there are researchers using the code now, that are part of a community we want to build connections with.

For those reasons we will proceed with a review.

I expect that we will find an editor by next Monday.

@skrawcz could you please go ahead and make a full submission issue? I will close this one once you have done so.
Thanks! Looking forward to the review 🚀
