### 2023-02-01 #74 @skrawcz
Okay I did https://github.com/pyOpenSci/software-submission/issues/80; still working on JOSS section, otherwise I think I filled it out appropriately.
